Mainly added the `cache_syncer` worker and removed that from exporter
The export worker should not sync the local/remote cache everytime before exporting, it is too much overhead.

So we introduced a cache syncer worker which is used to synchronize large files that do not change regularly. So for instance say that each job in the export/copyto/start phase requires a huge application to run the input files. That application will probably not change that often so we can cache it.

A caching worker job takes in your local cache dir `LocalSourceCacheFS()` and the `RemoteDestCacheFS()`. We then have two additional remote folder paths that get derived from the remote path, namely `_pending` and `_old`. When syncing does its job it will:
- check if syncing is required (whether local is in sync with remote/final dir)
- delete pending
- upload into pending
- delete old
- move final (which is previously uploaded) to old
- move pending to final