Add `Handlers` to CopyTo And CopyBack workers
These handlers can be used by consumers of this package to "mitigate" certain things like when failing to remove directories. It can be added to the consumer "todo admin task list" and the admin could even manually cleanup those or write a script to do so. The chances are high that if we try again later to remove the dir it will pass because a process could have been holding on to it that took a while to "die".